+++ src/App.js
@@ -3,6 +3,7 @@ import Home from '../src/Home/Home';
 import About from '../src/About/About';
 import Contact from './Contact/Contact';
 import Projects from './Projects/Projects';
+import Introduction from './Introduction/Introduction';
 
 // Install react-router-dom
 
@@ -10,7 +11,8 @@ function App() {
   return (
   <BrowserRouter>
     <Routes>
-      <Route path="/" element={<Home />} />
+      <Route path="/" element={<Introduction />} />
+      <Route path="home" element={<Home />} />
       <Route  path="about" element={<About />} />
       <Route path="projects" element={<Projects />} />
       <Route path="contact" element={<Contact />} />
