+++ src/Home/Home.css
@@ -20,7 +20,7 @@
     animation-duration: 1.5s;
     animation-iteration-count: 1;
     animation-direction: normal;
-    color: var(--h-color);
+    /* color: var(--h-color); */
 }
 
 @keyframes heading {
@@ -47,10 +47,25 @@
     animation-duration: 3s;
 }
 
-#myVideo {
-    position: fixed;
-    right:0;
-    bottom: 0;
-    min-width: 100%;
-    min-height: 100%;
+.video-wrapper {
+    position: relative;
+    overflow:hidden;
+    text-align: center;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    top: 0;
+    left:0;
+    width: 100%;
+    height: 100%;
+    z-index: -1;
+}
+
+video {
+  object-fit: cover;
+  height: 100%;
+  width: 100%;
+  position: fixed;
+  top: 0;
+  left: 0;
 }
