--- src/components/Introduction/Introduction.css
@@ -1,32 +1,0 @@
-#introduction {
-    padding: 5%;
-}
-
-.color-name {
-    background: linear-gradient(90deg, #00ffff,#0066ff);
-    -webkit-background-clip: text;
-    background-clip: text;
-    -webkit-text-fill-color: transparent;
-}
-
-div p {
-    font-size: 2.3rem;
-    font-weight: 700;
-    padding: 0 5%;
-    text-align: center;
-    text-wrap:wrap;
-}
-
-#clr-id {
-    background: linear-gradient(125deg, #00000070,#555555,#000);
-    -webkit-background-clip: text;
-    background-clip: text;
-    -webkit-text-fill-color: transparent;
-}
-
-#show-name {
-    background: linear-gradient(125deg, #cadd1bc7, rgba(226, 207, 207, 0.849), #e6e200fa);
-    -webkit-background-clip: text;
-    background-clip: text;
-    -webkit-text-fill-color: transparent;
-}
