--- src/components/Introduction/Introduction.js
@@ -1,11 +1,0 @@
-import './Introduction.css'
-
-function Introduction () {
-    return (
-        <div id="introduction">
-            <p>"Welcome to a visual journey where creativity meets craftsmanship, showcasing a portfolio that transcends boundaries and brings ideas to life with passion and precision."</p>
-        </div>
-    )
-}
-
-export default Introduction;
